#9 Using dynamic import for extension Upstream update to the latest 2022.02.xx version; Dependencies update.
