+++ js/extensions.js
@@ -6,11 +6,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { createPlugin } from "@mapstore/utils/PluginsUtils";
 import Extension from './extension/plugins/Extension';
 import { name } from '../config';
 
 
 export default {
-    [name]: createPlugin(name, Extension)
+    [name]: Extension
 };
